implement database for scribe (#125)
**Description**
Continuing on issue #114 and PR #116 / #118, this PR adds GORM database support for the generic event indexing scribe. It allows for two types of tables to be created, one for Logs and one for Receipts. Both tables have a "store" function as well as a "retrieve" function which is received by txHash.