add test contract and contract backfill to get logs (#137)
For the scribe, there is a need for backlogging logs from transactions that occurred. This is going to eventually pair with a live watcher that will add logs and receipts to the database in real time. Right now, this PR merely gets logs from a certain blockheight to another blockheight.